fix(mesi): move cache backends to sub-packages to fix Traefik Plugin Catalog
Move Memcached and Redis cache implementations from the mesi package to
separate sub-packages (mesi/cache_memcached and mesi/cache_redis). This
fixes issue #283 where Yaegi (Traefik's Go interpreter) fails to load
the plugin because it ignores //go:build tags and tries to compile all
files, including those that import gomemcache and go-redis packages not
available in Traefik's plugin catalog.

Key changes:
- Move cache_memcached.go to mesi/cache_memcached/cache_memcached.go
- Move cache_redis.go to mesi/cache_redis/cache_redis.go
- Update servers/caddy/mesi.go to use new sub-packages
- Update servers/traefik/cache_redis.go to use new sub-packages
- Remove old files from mesi/ package
- Update go.mod files

Fixes #283
