+++ mesi/cache_memcached/cache_memcached_test.go
@@ -1,6 +1,4 @@
-//go:build memcached
-
-package mesi
+package cache_memcached
 
 import (
 	"context"
