+++ servers/caddy/go.mod
@@ -9,7 +9,10 @@ require (
 	github.com/redis/go-redis/v9 v9.18.0
 )
 
-require github.com/crazy-goat/go-mesi v0.4.1
+require (
+	github.com/bradfitz/gomemcache v0.0.0-20250403215159-8d39553ac7cf
+	github.com/crazy-goat/go-mesi v0.4.1
+)
 
 replace github.com/crazy-goat/go-mesi => ../..
 
@@ -29,7 +32,6 @@ require (
 	github.com/antlr4-go/antlr/v4 v4.13.1 // indirect
 	github.com/aryann/difflib v0.0.0-20210328193216-ff5ff6dc229b // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/bradfitz/gomemcache v0.0.0-20250403215159-8d39553ac7cf // indirect
 	github.com/caddyserver/certmagic v0.25.2 // indirect
 	github.com/caddyserver/zerossl v0.1.5 // indirect
 	github.com/ccoveille/go-safecast/v2 v2.0.0 // indirect
